chore: change structure of selected item tooltip
