+++ src/components/ComboBox/MultiSelectedItemTooltip.tsx
@@ -1,0 +1,19 @@
+import React, { ReactNode, VFC } from 'react'
+
+import { Tooltip } from '../Tooltip'
+
+type Props = {
+  needsTooltip: boolean
+  text: string
+  children: ReactNode
+}
+
+export const MultiSelectedItemTooltip: VFC<Props> = ({ needsTooltip, text, children }) => {
+  return needsTooltip ? (
+    <Tooltip message={text} multiLine>
+      {children}
+    </Tooltip>
+  ) : (
+    <>{children}</>
+  )
+}
